bert_pos

+++ app/run_bert_pos.py
@@ -144,7 +144,7 @@ def main(args):
                 # logger.info(('size pos1: {}, pos2: {}, real_feature: {}'.format(pos1.size(), pos2.size(), real_feature.size())))
                 out = pos(pos1, pos2, real_feature)
 
-                loss = adversarial_loss(out, y.long())
+                loss = adversarial_loss(out, y.float())
                 loss.backward()
                 total_loss += loss.detach()
